fixed calculateHelper
