+++ hw02/src/main/java/ru/sberbank/homework/kashin/main/model/numbers/BinaryNumber.java
@@ -3,8 +3,8 @@ package ru.sberbank.homework.kashin.main.model.numbers;
 import ru.sberbank.homework.kashin.main.exception.WrongExpression;
 import ru.sberbank.homework.kashin.main.model.Number;
 
-import static ru.sberbank.homework.kashin.main.util.CalculateHelper.getOriginalLiterals;
-import static ru.sberbank.homework.kashin.main.util.CalculateHelper.setPreResult;
+import static ru.sberbank.homework.kashin.main.calculator.ExpressionCalculator.originalLiterals;
+import static ru.sberbank.homework.kashin.main.calculator.ExpressionCalculator.setPreResult;
 
 public class BinaryNumber extends Number {
     @Override
@@ -17,7 +17,7 @@ public class BinaryNumber extends Number {
             }
         } catch (Exception e) {
             setPreResult(null);
-            throw new WrongExpression(String.format("error > %s", getOriginalLiterals().get(item)));
+            throw new WrongExpression(String.format("error > %s", originalLiterals.get(item)));
         }
     }
